requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.12
matplotlib>=3.3.4
numpy>=1.21.0
tqdm>=4.62.0
scikit-learn>=1.0.0
Pillow>=8.3.1
seaborn>=0.11.2
streamlit>=1.12.0
opencv-python>=4.5.0
gradio>=3.0.0
pandas>=1.3.0